Fixed a segmentation fault that would happen when a non existent unit id was attempted to be deleted.
* Also added debug to makefile for lldb: clean and run 'make debug' to build with symbols
Bug repro:
spawn 0 0 -1 1
spawn 0 -1 0 1
colonize 0 0 0 0
units
<seg fault>

The units query would for_each_unit and there was a nullptr unit in the map.
This happened when a unit that didn't exist was attempted to be deleted, from step colonize 0 0 0 0.
That colonize step attempted to get unit with id 0 to delete. Since a raw index into the map was used
an invalid entry was being put into the units map.

Resolution:
Used get_unit(id) instead of units[id]